+++ webui/src/store/actionTypes.js
@@ -187,3 +187,4 @@ export const RENAME_FOLDER = 'RENAME_FOLDER'
 // export const deleteFolder = types(DELETE_FOLDER)
 // export const renameFile = types(RENAME_FILE)
 // export const renameFolder = types(RENAME_FOLDER)
+