+++ webui/src/store/request/queries.graphql.js
@@ -524,3 +524,4 @@ export const getFilesQuery = gql`
     }
   }
 `;
+